fix: jumping text due to font size being calculated incorrectly

+++ src/element/textElement.ts
@@ -280,7 +280,7 @@ export const measureText = (
     // lines would be stripped from computation
     .map((x) => x || " ")
     .join("\n");
-  const fontSize = parseInt(font);
+  const fontSize = parseFloat(font);
   const height = getTextHeight(text, fontSize, lineHeight);
   const width = getTextWidth(text, font);
